Being more specific about the exceptions
