+++ MSBLOC.Core/Model/ProjectDetails.cs
@@ -22,12 +22,11 @@ namespace MSBLOC.Core.Model
 
             if (!projectFile.IsSubPathOf(cloneRoot))
             {
-                throw new ProjectDetailsException("Project file path is not a subpath of clone root");
+                throw new ProjectDetailsException($"Project file path \"{projectFile}\" is not a subpath of \"{cloneRoot}\"");
             }
 
             ProjectDirectory = Path.GetDirectoryName(projectFile)
-                               ?? throw new InvalidOperationException(
-                                   "Path.GetDirectoryName(startedEventArgs.ProjectFile) is null");
+                               ?? throw new ProjectDetailsException("Path.GetDirectoryName(projectFile) is null");
 
             _paths = new Dictionary<string, string>();
         }
@@ -44,7 +43,7 @@ namespace MSBLOC.Core.Model
         {
             if (_paths.ContainsKey(itemProjectPath))
             {
-                throw new ArgumentException(nameof(itemProjectPath));
+                throw new ProjectDetailsException($"Item \"{itemProjectPath}\" already exists");
             }
 
             _paths[itemProjectPath] = GetClonePath(itemProjectPath);
